requirements.txt
# Django
django
Pillow
pylint-django

# Postgresql
psycopg2-binary

# Dot enviroment
python-dotenv

# Google API
google-auth
google-auth-oauthlib
google-auth-httplib2
